requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
black==22.6.0
certifi==2022.6.15
cfgv==3.3.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
Deprecated==1.2.13
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.6
django-cors-headers==3.13.0
django-redis==5.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
filelock==3.7.1
gunicorn==20.1.0
identify==2.5.1
idna==3.3
mypy-extensions==0.4.3
nodeenv==1.7.0
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pre-commit==2.19.0
psycopg2-binary==2.9.3
PyJWT==2.4.0
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
redis==4.3.4
regex==2022.6.2
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.9
virtualenv==20.15.1
whitenoise==6.2.0
wincertstore==0.2
wrapt==1.14.1
